# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Team Task Manager API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  admin: "admin",
  member: "member",
} as const;

export interface User {
  id: number;
  email: string;
  name: string;
  role: UserRole;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  message: string;
}

export type SignupBodyRole =
  (typeof SignupBodyRole)[keyof typeof SignupBodyRole];

export const SignupBodyRole = {
  admin: "admin",
  member: "member",
} as const;

export interface SignupBody {
  email: string;
  password: string;
  name: string;
  role?: SignupBodyRole;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface Project {
  id: number;
  name: string;
  description?: string | null;
  createdById: number;
  createdAt: string;
  memberCount: number;
  taskCount: number;
}

export type TaskStatus = (typeof TaskStatus)[keyof typeof TaskStatus];

export const TaskStatus = {
  todo: "todo",
  "in-progress": "in-progress",
  done: "done",
} as const;

export interface Task {
  id: number;
  title: string;
  description?: string | null;
  status: TaskStatus;
  projectId: number;
  assignedToId?: number | null;
  assignedTo?: User | null;
  dueDate?: string | null;
  createdAt: string;
}

export interface ProjectDetail {
  id: number;
  name: string;
  description?: string | null;
  createdById: number;
  createdAt: string;
  members: User[];
  tasks: Task[];
}

export interface CreateProjectBody {
  name: string;
  description?: string;
}

export interface UpdateProjectBody {
  name?: string;
  description?: string;
}

export interface AddMemberBody {
  userId: number;
}

export type CreateTaskBodyStatus =
  (typeof CreateTaskBodyStatus)[keyof typeof CreateTaskBodyStatus];

export const CreateTaskBodyStatus = {
  todo: "todo",
  "in-progress": "in-progress",
  done: "done",
} as const;

export interface CreateTaskBody {
  title: string;
  description?: string;
  assignedToId?: number;
  dueDate?: string;
  status?: CreateTaskBodyStatus;
}

export type UpdateTaskBodyStatus =
  (typeof UpdateTaskBodyStatus)[keyof typeof UpdateTaskBodyStatus];

export const UpdateTaskBodyStatus = {
  todo: "todo",
  "in-progress": "in-progress",
  done: "done",
} as const;

export interface UpdateTaskBody {
  title?: string;
  description?: string;
  assignedToId?: number;
  dueDate?: string;
  status?: UpdateTaskBodyStatus;
}

export type DashboardStatsTasksByStatus = {
  todo: number;
  "in-progress": number;
  done: number;
};

export interface DashboardStats {
  totalTasks: number;
  tasksByStatus: DashboardStatsTasksByStatus;
  overdueTasks: number;
  totalProjects: number;
  myTasks: number;
}