# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Team Task Manager API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: projects
    description: Project management
  - name: tasks
    description: Task management
  - name: dashboard
    description: Dashboard statistics
  - name: users
    description: User management
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/signup:
    post:
      operationId: signup
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects for current user
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"

    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectBody"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"

  /projects/{projectId}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get project by ID
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      operationId: updateProject
      tags: [projects]
      summary: Update project (admin only)
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectBody"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"

    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete project (admin only)
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /projects/{projectId}/members:
    post:
      operationId: addProjectMember
      tags: [projects]
      summary: Add a member to a project (admin only)
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddMemberBody"
      responses:
        "201":
          description: Member added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /projects/{projectId}/members/{userId}:
    delete:
      operationId: removeProjectMember
      tags: [projects]
      summary: Remove a member from a project (admin only)
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Member removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /projects/{projectId}/tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"

    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task in a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskBody"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"

  /tasks/{taskId}:
    put:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskBody"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"

    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      summary: Get dashboard statistics
      responses:
        "200":
          description: Dashboard statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"

  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (admin only)
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, member]
        createdAt:
          type: string
      required:
        - id
        - email
        - name
        - role
        - createdAt

    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        message:
          type: string
      required:
        - user
        - message

    SignupBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, member]
      required:
        - email
        - password
        - name

    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password

    Project:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        createdById:
          type: integer
        createdAt:
          type: string
        memberCount:
          type: integer
        taskCount:
          type: integer
      required:
        - id
        - name
        - createdById
        - createdAt
        - memberCount
        - taskCount

    ProjectDetail:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        createdById:
          type: integer
        createdAt:
          type: string
        members:
          type: array
          items:
            $ref: "#/components/schemas/User"
        tasks:
          type: array
          items:
            $ref: "#/components/schemas/Task"
      required:
        - id
        - name
        - createdById
        - createdAt
        - members
        - tasks

    CreateProjectBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
      required:
        - name

    UpdateProjectBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string

    AddMemberBody:
      type: object
      properties:
        userId:
          type: integer
      required:
        - userId

    Task:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [todo, in-progress, done]
        projectId:
          type: integer
        assignedToId:
          type: integer
          nullable: true
        assignedTo:
          $ref: "#/components/schemas/User"
          nullable: true
        dueDate:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - title
        - status
        - projectId
        - createdAt

    CreateTaskBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        assignedToId:
          type: integer
        dueDate:
          type: string
        status:
          type: string
          enum: [todo, in-progress, done]
      required:
        - title

    UpdateTaskBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        assignedToId:
          type: integer
        dueDate:
          type: string
        status:
          type: string
          enum: [todo, in-progress, done]

    DashboardStats:
      type: object
      properties:
        totalTasks:
          type: integer
        tasksByStatus:
          type: object
          properties:
            todo:
              type: integer
            in-progress:
              type: integer
            done:
              type: integer
          required:
            - todo
            - in-progress
            - done
        overdueTasks:
          type: integer
        totalProjects:
          type: integer
        myTasks:
          type: integer
      required:
        - totalTasks
        - tasksByStatus
        - overdueTasks
        - totalProjects
        - myTasks