# lib/api-zod/src/generated/types/task.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Team Task Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { TaskStatus } from "./taskStatus";
import type { User } from "./user";

export interface Task {
  id: number;
  title: string;
  description?: string | null;
  status: TaskStatus;
  projectId: number;
  assignedToId?: number | null;
  assignedTo?: User | null;
  dueDate?: string | null;
  createdAt: string;
}